implemented label getter
