+++ packages/graph-plugins/src/plugins/node-label/constants.ts
@@ -1,0 +1,1 @@
+export const UPPERCASE_ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'.split('');
